set up recycler view with paging

+++ app/src/main/java/jiglionero/android/app/putonpompom/domain/forecast/Main.kt
@@ -6,7 +6,7 @@ data class Main(
     val pressure: Double = 0.0,
     val sea_level: Double = 0.0,
     val temp: Double = 0.0,
-    val temp_kf: Int = 0,
+    val temp_kf: Double = 0.0,
     val temp_max: Double = 0.0,
     val temp_min: Double = 0.0
 )
